+++ projects/angular-cesium/src/lib/angular-cesium-widgets/services/entity-editors/polygons-editor/polygons-editor.service.ts
@@ -1,4 +1,4 @@
-import { debounceTime, publish, tap } from 'rxjs/operators';
+import { publish, tap } from 'rxjs/operators';
 import { Injectable } from '@angular/core';
 import { CesiumService } from '../../../../angular-cesium/services/cesium/cesium.service';
 import { MapEventsManagerService } from '../../../../angular-cesium/services/map-events-mananger/map-events-manager';
@@ -120,17 +120,26 @@ export class PolygonsEditorService {
     return this.updatePublisher;
   }
 
-  screenToPosition(cartesian2, clampHeightTo3D: boolean) {
+  screenToPosition(cartesian2, clampHeightTo3D: boolean, isTerrain?) {
     const cartesian3 = this.coordinateConverter.screenToCartesian3(cartesian2);
+
     // If cartesian3 is undefined then the point inst on the globe
     if (clampHeightTo3D && cartesian3) {
-      const cartesian3PickPosition = this.cesiumScene.pickPosition(cartesian2);
-      const latLon = CoordinateConverter.cartesian3ToLatLon(cartesian3PickPosition);
-      if (latLon.height < 0) {// means nothing picked -> Validate it
+      const globePositionPick = () => {
         const ray = this.cameraService.getCamera().getPickRay(cartesian2);
         return this.cesiumScene.globe.pick(ray, this.cesiumScene);
+      };
+
+      if (isTerrain) {
+        return globePositionPick();
+      } else {
+        const cartesian3PickPosition = this.cesiumScene.pickPosition(cartesian2);
+        const latLon = CoordinateConverter.cartesian3ToLatLon(cartesian3PickPosition);
+        if (latLon.height < 0) {// means nothing picked -> Validate it
+          return globePositionPick();
+        }
+        return this.cesiumScene.clampToHeight(cartesian3PickPosition);
       }
-      return this.cesiumScene.clampToHeight(cartesian3PickPosition);
     }
 
     return cartesian3;
@@ -178,7 +187,7 @@ export class PolygonsEditorService {
     const editorObservable = this.createEditorObservable(clientEditSubject, id);
 
     mouseMoveRegistration.subscribe(({ movement: { endPosition } }) => {
-      const position = this.screenToPosition(endPosition, options.clampHeightTo3D);
+      const position = this.screenToPosition(endPosition, polygonOptions.clampHeightTo3D, polygonOptions.clampHeightTo3DOptions.clampToTerrain);
 
       if (position) {
         this.updateSubject.next({
@@ -195,7 +204,7 @@ export class PolygonsEditorService {
       if (finishedCreate) {
         return;
       }
-      const position = this.screenToPosition(endPosition, options.clampHeightTo3D);
+      const position = this.screenToPosition(endPosition, polygonOptions.clampHeightTo3D, polygonOptions.clampHeightTo3DOptions.clampToTerrain);
       if (!position) {
         return;
       }
@@ -233,7 +242,7 @@ export class PolygonsEditorService {
 
 
     addLastPointRegistration.subscribe(({ movement: { endPosition } }) => {
-      const position = this.screenToPosition(endPosition, options.clampHeightTo3D);
+      const position = this.screenToPosition(endPosition, polygonOptions.clampHeightTo3D, polygonOptions.clampHeightTo3DOptions.clampToTerrain);
       if (!position) {
         return;
       }
@@ -360,7 +369,7 @@ export class PolygonsEditorService {
     pointDragRegistration.pipe(
       tap(({ movement: { drop } }) => this.cameraService.enableInputs(drop)))
       .subscribe(({ movement: { endPosition, drop }, entities }) => {
-        const position = this.screenToPosition(endPosition, options.clampHeightTo3D);
+        const position = this.screenToPosition(endPosition, options.clampHeightTo3D, options.clampHeightTo3DOptions.clampToTerrain);
         if (!position) {
           return;
         }
@@ -472,7 +481,7 @@ export class PolygonsEditorService {
 
       polygonOptions.allowDrag = false;
       polygonOptions.polylineProps.clampToGround = true;
-      polygonOptions.pointProps.heightReference =  polygonOptions.clampHeightTo3DOptions.clampToTerrain ?
+      polygonOptions.pointProps.heightReference = polygonOptions.clampHeightTo3DOptions.clampToTerrain ?
         Cesium.HeightReference.CLAMP_TO_GROUND : Cesium.HeightReference.RELATIVE_TO_GROUND;
       polygonOptions.pointProps.disableDepthTestDistance = Number.POSITIVE_INFINITY;
     }